// src/core/api/generated/spring/schemas/TenantSetupProgressDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Crm Backend API
 * Crm Backend API documentation
 * OpenAPI spec version: 0.0.1
 */
import type { TenantSetupProgressDTOStatus } from './TenantSetupProgressDTOStatus';
import type { TenantSetupProgressDTOCurrentStep } from './TenantSetupProgressDTOCurrentStep';

export interface TenantSetupProgressDTO {
  tenantId?: string;
  tenantName?: string;
  status?: TenantSetupProgressDTOStatus;
  currentStep?: TenantSetupProgressDTOCurrentStep;
  progressPercentage?: number;
  message?: string;
  errorMessage?: string;
  startTime?: string;
  endTime?: string;
  estimatedTimeRemainingSeconds?: number;
}